// testservices/models/VehicleCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type VehicleCreate = {
    vehicle_name: string;
    launch_flags?: (string | null);
    assigned_bundle_name?: (string | null);
    assigned_group_name?: (string | null);
};